move jdk14 tests and add with build-helper-maven-plugin
