+++ pom.xml
@@ -94,6 +94,12 @@
 		<pluginManagement>
 			<plugins>
 				<plugin>
+					<groupId>org.codehaus.mojo</groupId>
+					<artifactId>build-helper-maven-plugin</artifactId>
+					<version>3.2.0</version>
+				</plugin>
+
+				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
 					<version>3.8.1</version>
@@ -195,7 +201,6 @@
 			</build>
 		</profile>
 		<!-- Exclude any tests that require JDK 11+ from test run -->
-		<!-- Exclude any tests that require JDK 14+ from compilation and test run -->
 		<profile>
 			<id>until-jdk11</id>
 			<activation>
@@ -209,30 +214,11 @@
 					<plugins>
 						<plugin>
 							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-compiler-plugin</artifactId>
-							<executions>
-								<execution>
-									<id>default-testCompile</id>
-									<phase>test-compile</phase>
-									<goals>
-										<goal>testCompile</goal>
-									</goals>
-									<configuration>
-										<testExcludes>
-											<testExclude>jdk14/**</testExclude>
-										</testExcludes>
-									</configuration>
-								</execution>
-							</executions>
-						</plugin>
-						<plugin>
-							<groupId>org.apache.maven.plugins</groupId>
 							<artifactId>maven-surefire-plugin</artifactId>
 							<version>2.22.0</version>
 							<configuration>
 								<excludes>
 									<exclude>jdk11/**</exclude>
-									<exclude>jdk14/**</exclude>
 								</excludes>
 							</configuration>
 						</plugin>
@@ -240,73 +226,48 @@
 				</pluginManagement>
 			</build>
 		</profile>
-		<!-- Exclude any tests that require JDK 14+ from compilation and test run-->
-		<profile>
-			<id>jdk11ge</id>
-			<activation>
-				<jdk>[11,14)</jdk>
-			</activation>
-			<build>
-				<pluginManagement>
-					<plugins>
-						<plugin>
-							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-compiler-plugin</artifactId>
-							<executions>
-								<execution>
-									<id>default-testCompile</id>
-									<phase>test-compile</phase>
-									<goals>
-										<goal>testCompile</goal>
-									</goals>
-									<configuration>
-										<testExcludes>
-											<testExclude>jdk14/**</testExclude>
-										</testExcludes>
-									</configuration>
-								</execution>
-							</executions>
-						</plugin>
-						<plugin>
-							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-surefire-plugin</artifactId>
-							<configuration>
-								<excludes>
-									<exclude>jdk14/**</exclude>
-								</excludes>
-							</configuration>
-						</plugin>
-					</plugins>
-				</pluginManagement>
-			</build>
-		</profile>
-		<!-- Run all tests with JDK 14+ if available -->
+		<!-- Include any tests that require JDK14+ -->
 		<profile>
 			<id>jdk14ge</id>
 			<activation>
-				<jdk>[14,)</jdk>
+				<jdk>[14,</jdk>
 			</activation>
 			<build>
 				<pluginManagement>
 					<plugins>
 						<plugin>
-							<groupId>org.apache.maven.plugins</groupId>
-							<artifactId>maven-compiler-plugin</artifactId>
+							<groupId>org.codehaus.mojo</groupId>
+							<artifactId>build-helper-maven-plugin</artifactId>
 							<executions>
 								<execution>
-									<id>default-testCompile</id>
-									<phase>test-compile</phase>
+									<id>add-test-source</id>
+									<phase>generate-test-sources</phase>
 									<goals>
-										<goal>testCompile</goal>
+										<goal>add-test-source</goal>
 									</goals>
 									<configuration>
-										<release>${java.vm.specification.version}</release>
-										<compilerArgs>
-											<arg>--enable-preview</arg>
-										</compilerArgs>
+										<sources>
+											<source>test-jdk14</source>
+										</sources>
 									</configuration>
 								</execution>
 							</executions>
+						</plugin>
+						<plugin>
+							<groupId>org.apache.maven.plugins</groupId>
+							<artifactId>maven-compiler-plugin</artifactId>
+							<configuration>
+								<optimize>true</optimize>
+								<!-- Enable Java 14+ for all sources so that Intellij picks the right language level -->
+								<source>${java.vm.specification.version}</source>
+								<target>${java.vm.specification.version}</target>
+								<compilerArgs>
+									<arg>-parameters</arg>
+									<arg>--enable-preview</arg>
+									<arg>--add-exports=java.base/jdk.internal.ref=ALL-UNNAMED</arg>
+									<arg>--add-exports=java.base/sun.nio.ch=ALL-UNNAMED</arg>
+								</compilerArgs>
+							</configuration>
 						</plugin>
 						<plugin>
 							<groupId>org.apache.maven.plugins</groupId>
