+++ test-jdk14/com/esotericsoftware/kryo/TestDataJava14.java
@@ -18,7 +18,7 @@
  * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
-package jdk14;
+package com.esotericsoftware.kryo;
 
 import com.esotericsoftware.kryo.SerializationCompatTestData;
