+++ test/com/esotericsoftware/kryo/SerializationCompatTest.java
@@ -77,10 +77,14 @@ public class SerializationCompatTest extends KryoTestCase {
 
 	private static final int JAVA_VERSION;
 	static {
-		// java.version is e.g. 1.8.0 or 9.0.4
+		// java.version is e.g. "1.8.0", "9.0.4", or "14"
 		String[] strVersions = System.getProperty("java.version").split("\\.");
-		int[] versions = new int[] {parseInt(strVersions[0]), parseInt(strVersions[1])};
-		JAVA_VERSION = versions[0] > 1 ? versions[0] : versions[1];
+		if (strVersions.length == 1) {
+			JAVA_VERSION = parseInt(strVersions[0]);
+		} else {
+			int[] versions = new int[] {parseInt(strVersions[0]), parseInt(strVersions[1])};
+			JAVA_VERSION = versions[0] > 1 ? versions[0] : versions[1];
+		}
 	}
 	private static final int EXPECTED_DEFAULT_SERIALIZER_COUNT = JAVA_VERSION < 11
 			? 57 : JAVA_VERSION < 14 ? 67 : 68;  // Also change Kryo#defaultSerializers.
@@ -90,12 +94,12 @@ public class SerializationCompatTest extends KryoTestCase {
 		TEST_DATAS.add(new TestDataDescription<>(new TestData(), 1940, 1958));
 		if (JAVA_VERSION >= 8) TEST_DATAS.add(new TestDataDescription<>(new TestDataJava8(), 2098, 2116));
 		if (JAVA_VERSION >= 11) TEST_DATAS.add(new TestDataDescription<>(new TestDataJava11(), 2210, 2238));
-		if (JAVA_VERSION >= 14) TEST_DATAS.add(new TestDataDescription<>(createTestDataJava14(), 1928, 1946));
+		if (JAVA_VERSION >= 14) TEST_DATAS.add(new TestDataDescription<>(createTestDataJava14(), 1948, 1966));
 	};
 
 	private static TestData createTestDataJava14() {
 		try {
-			return (TestData) Class.forName("jdk14.TestDataJava14").getConstructor().newInstance();
+			return (TestData) Class.forName("com.esotericsoftware.kryo.TestDataJava14").getConstructor().newInstance();
 		} catch (ReflectiveOperationException e) {
 			throw new RuntimeException("TestDataJava14 could not be instantiated", e);
 		}
